F: Improve syntax errors accuracy
