+++ rsconcept/domain/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@rsconcept/domain",
-  "version": "1.7.0",
+  "version": "1.7.2",
   "description": "Shared TypeScript domain for RS language, RSForm, RSModel, RSEngine, OSS, and concept-text helpers.",
   "license": "MIT",
   "author": "IRBorisov",
