+++ rsconcept/frontend/package.json
@@ -30,7 +30,7 @@
     "@lezer/highlight": "^1.2.3",
     "@lezer/lr": "^1.4.10",
     "@react-pdf/renderer": "^4.5.1",
-    "@rsconcept/domain": "^1.7.0",
+    "@rsconcept/domain": "^1.7.2",
     "@sentry/react": "^10.62.0",
     "@tanstack/react-form": "^1.33.0",
     "@tanstack/react-query": "^5.101.2",
