+++ rsconcept/rstool/docs/DIAGNOSTICS.md
@@ -37,24 +37,29 @@
 
 ## Parser
 
-| Код      | Имя                       | `params`                  | Исправление                                                                           |
-| :------- | :------------------------ | :------------------------ | :------------------------------------------------------------------------------------ |
-| `0x8400` | `unknownSyntax`           | —                         | Неопределённая синтаксическая ошибка: проверь операторы, скобки, идентификаторы.      |
-| `0x8401` | `forbiddenCharacter`      | `[character]`             | Символ не входит в язык родов структур (напр. `^` для степени). Удали или замени его. |
-| `0x8409` | `bracketMismatch`         | `[expected, actual]`      | Закрой ожидаемой скобкой `expected` вместо `actual`.                                  |
-| `0x840A` | `doubleParenthesis`       | —                         | Убери лишнюю внешнюю пару `(( … ))`.                                                  |
-| `0x840B` | `missingOpenBracket`      | `[bracket]`               | Непарная закрывающая `bracket`: убери её или добавь открывающую.                      |
-| `0x840C` | `missingCloseBracket`     | `[bracket]`               | Непарная открывающая `bracket`: убери её или добавь закрывающую.                      |
-| `0x8415` | `expectedLocal`           | —                         | Ожидалось имя локальной переменной.                                                   |
-| `0x8416` | `expectedType`            | `[expectedClass, actual]` | Несоответствие класса выражения ожидаемому.                                           |
-| `0x8417` | `invalidFilterSyntax`     | —                         | Синтаксис `Fi`: `Fi1[D1](S1)`, `Fi1,2[D1,D2](S1)` или `Fi1,2[D3](S1)`.                |
-| `0x8418` | `expectedFunctionBody`    | —                         | Нужно тело функции после `[параметры]`.                                               |
-| `0x8419` | `expectedExpressionBody`  | —                         | Нужно теоретико-множественное тело функции (ТФ).                                      |
-| `0x841A` | `expectedLogicBody`       | —                         | Нужно логическое тело функции.                                                        |
-| `0x841B` | `expectedQuantifierBody`  | —                         | Нужно логическое тело после `∀(∃)…∈домен`.                                            |
-| `0x841C` | `expectedDeclarativeBody` | —                         | Допиши декларативное `D{…∈… \| …}`.                                                   |
-| `0x841D` | `expectedImperativeBody`  | —                         | Допиши императивное `I{… \| …}`.                                                      |
-| `0x841E` | `expectedRecursiveBody`   | —                         | Допиши рекурсивное `R{…:=… \| … \| …}`.                                               |
+| Код      | Имя                        | `params`                  | Исправление                                                                           |
+| :------- | :------------------------- | :------------------------ | :------------------------------------------------------------------------------------ |
+| `0x8400` | `unknownSyntax`            | —                         | Неопределённая синтаксическая ошибка: проверь операторы, скобки, идентификаторы.      |
+| `0x8401` | `forbiddenCharacter`       | `[character]`             | Символ не входит в язык родов структур (напр. `^` для степени). Удали или замени его. |
+| `0x8409` | `bracketMismatch`          | `[expected, actual]`      | Закрой ожидаемой скобкой `expected` вместо `actual`.                                  |
+| `0x840A` | `doubleParenthesis`        | —                         | Убери лишнюю внешнюю пару `(( … ))`.                                                  |
+| `0x840B` | `missingOpenBracket`       | `[bracket]`               | Непарная закрывающая `bracket`: убери её или добавь открывающую.                      |
+| `0x840C` | `missingCloseBracket`      | `[bracket]`               | Непарная открывающая `bracket`: убери её или добавь закрывающую.                      |
+| `0x8415` | `expectedLocal`            | —                         | Ожидалось имя локальной переменной.                                                   |
+| `0x8416` | `expectedType`             | `[expectedClass, actual]` | Несоответствие класса выражения ожидаемому.                                           |
+| `0x8417` | `invalidFilterSyntax`      | —                         | Синтаксис `Fi`: `Fi1[D1](S1)`, `Fi1,2[D1,D2](S1)` или `Fi1,2[D3](S1)`.                |
+| `0x8418` | `expectedFunctionBody`     | —                         | Нужно тело функции после `[параметры]`.                                               |
+| `0x8419` | `expectedExpressionBody`   | —                         | Нужно теоретико-множественное тело функции (ТФ).                                      |
+| `0x841A` | `expectedLogicBody`        | —                         | Нужно логическое тело функции.                                                        |
+| `0x841B` | `expectedQuantifierBody`   | —                         | Нужно логическое тело после `∀(∃)…∈домен`.                                            |
+| `0x841C` | `expectedDeclarativeBody`  | —                         | Допиши декларативное `D{…∈… \| …}`.                                                   |
+| `0x841D` | `expectedImperativeBody`   | —                         | Допиши императивное `I{… \| …}`.                                                      |
+| `0x841E` | `expectedRecursiveBody`    | —                         | Допиши рекурсивное `R{…:=… \| … \| …}`.                                               |
+| `0x841F` | `expectedQuantifierDomain` | —                         | После переменной квантора нужны `∈` и домен.                                          |
+| `0x8420` | `expectedRightOperand`     | —                         | Нужен правый операнд после оператора.                                                 |
+| `0x8421` | `expectedUnaryOperand`     | —                         | Нужно выражение после `¬`.                                                            |
+| `0x8422` | `globalFuncParenCall`      | `[name]`                  | Вызов `F`/`P` только через `[…]`, не `(…)`.                                           |
+| `0x8423` | `expectedArgument`         | —                         | Нужен аргумент в списке `[…]` у F/P.                                                  |
 
 ## Предупреждения
